refactor(agentic-ai): move response types to shared package and introduce AiAgentTaskResponse
Move AiAgentSubProcessResponse from jobworker subpackage to the shared
aiagent package. Introduce AiAgentTaskResponse implementing
StandardConnectorResponse for the outbound task connector flavor,
replacing the raw AgentResponse return type.

This aligns both connector flavors to return proper ConnectorResponse
implementations, preparing for the generic type bound tightening in the
next commit.
