+++ connectors/agentic-ai/src/main/java/io/camunda/connector/agenticai/aiagent/AiAgentSubProcessResponse.java
@@ -4,7 +4,7 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.agenticai.aiagent.jobworker;
+package io.camunda.connector.agenticai.aiagent;
 
 import io.camunda.connector.agenticai.aiagent.model.AgentResponse;
 import io.camunda.connector.agenticai.model.AgenticAiRecord;
