+++ connectors/agentic-ai/src/main/java/io/camunda/connector/agenticai/aiagent/agent/OutboundConnectorAgentRequestHandler.java
@@ -8,6 +8,7 @@ package io.camunda.connector.agenticai.aiagent.agent;
 
 import static io.camunda.connector.agenticai.aiagent.agent.AgentErrorCodes.ERROR_CODE_NO_USER_MESSAGE_CONTENT;
 
+import io.camunda.connector.agenticai.aiagent.AiAgentTaskResponse;
 import io.camunda.connector.agenticai.aiagent.framework.AiFrameworkAdapter;
 import io.camunda.connector.agenticai.aiagent.memory.conversation.ConversationStoreRegistry;
 import io.camunda.connector.agenticai.aiagent.model.AgentContext;
@@ -20,7 +21,7 @@ import java.util.List;
 import org.springframework.util.CollectionUtils;
 
 public class OutboundConnectorAgentRequestHandler
-    extends BaseAgentRequestHandler<OutboundConnectorAgentExecutionContext, AgentResponse> {
+    extends BaseAgentRequestHandler<OutboundConnectorAgentExecutionContext, AiAgentTaskResponse> {
 
   public OutboundConnectorAgentRequestHandler(
       AgentInitializer agentInitializer,
@@ -55,8 +56,8 @@ public class OutboundConnectorAgentRequestHandler
   }
 
   @Override
-  public AgentResponse completeJob(
+  public AiAgentTaskResponse completeJob(
       OutboundConnectorAgentExecutionContext executionContext, AgentResponse agentResponse) {
-    return agentResponse;
+    return new AiAgentTaskResponse(agentResponse);
   }
 }
